lab02/application/repository/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
psycopg2-binary==2.9.10